// src/shared/types/types.ts
export interface IProduct {
  id: number;
  title: string;
  description: string;
  price: number;
  discountPercentage: number;
  rating: number;
  stock: number;
  brand: string;
  category: string;
  thumbnail: string;
  images: Array<string>;
}

export interface IPayloadApi {
  products: IProduct[];
  total: number;
  skip: number;
  limit: number;
}